requirements.txt
boto3
Pillow
pymupdf
pytesseract
Spire.Doc